+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Removed
+* Error::InvalidTid (unused)
+
 ## [0.11.0] - 2026-03-10
 ### Removed
 * Moved `geo` and `mapgrid` modules to squarepeg crate